config: udpate eslint to latest version with Flat Config
I migrated the ESLint Config to the latest version, which deprecates the previous way of specifying and passes to a new Flat Config API.

I also had to upgrade the eslint-astro version. These are specified with links below.

Note: Migrating the config was painfully difficult, as there were a lot of major changes. Even now, there are still some issues, like:

1. When running the lint command, `pnpm run lint`, it displays errors from the `.astro` folder, which shouldn't be considerated, as it's in the `ignores` property from the config file.
2. There are some duplicate warnings/errors which belong to different configs. At least that's my guess. I must check how to avoid repetition, as it's annoying.

- https://ota-meshi.github.io/eslint-plugin-astro/user-guide/
- https://eslint.org/docs/latest/use/migrate-to-9.0.0
- https://eslint.org/docs/latest/use/configure/migration-guide
- https://typescript-eslint.io/getting-started/
-